% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_stat.R
\name{summarize_stat}
\alias{summarize_stat}
\title{Summary Statistics of a numeric column grouped by a category column in a dataset}
\usage{
summarize_stat(dataset, category, column)
}
\arguments{
\item{dataset}{The dataset.}

\item{category}{The category which can be divided into several types, and then we will summarize a numeric column grouped by this category.}

\item{column}{The numeric column for computing summary statistics.}
}
\value{
A tibble of the summary statistics.
}
\description{
Summary Statistics of a numeric column grouped by a category column in a dataset
}
\examples{
summarize_stat(datateachr::apt_buildings, property_type, no_of_storeys)
summarize_stat(datateachr::apt_buildings, air_conditioning, year_built)
summarize_stat(datateachr::apt_buildings, window_type, year_built)

}
